backend/requirements.txt
selenium>=4.0
opencc-python-reimplemented>=0.1.7
